Move comment related functions to comments.py and get rid of magic strings
