+++ src/black/nodes.py
@@ -838,29 +838,6 @@ def is_async_stmt_or_funcdef(leaf: Leaf) -> bool:
     )
 
 
-def is_type_comment(leaf: Leaf) -> bool:
-    """Return True if the given leaf is a type comment. This function should only
-    be used for general type comments (excluding ignore annotations, which should
-    use `is_type_ignore_comment`). Note that general type comments are no longer
-    used in modern version of Python, this function may be deprecated in the future."""
-    t = leaf.type
-    v = leaf.value
-    return t in {token.COMMENT, STANDALONE_COMMENT} and v.startswith("# type:")
-
-
-def is_type_ignore_comment(leaf: Leaf) -> bool:
-    """Return True if the given leaf is a type comment with ignore annotation."""
-    t = leaf.type
-    v = leaf.value
-    return t in {token.COMMENT, STANDALONE_COMMENT} and is_type_ignore_comment_string(v)
-
-
-def is_type_ignore_comment_string(value: str) -> bool:
-    """Return True if the given string match with type comment with
-    ignore annotation."""
-    return value.startswith("# type: ignore")
-
-
 def wrap_in_parentheses(parent: Node, child: LN, *, visible: bool = True) -> None:
     """Wrap `child` in parentheses.
